+++ packages/ui/BodyText/BodyText.js
@@ -6,8 +6,11 @@
  */
 
 import kind from '@enact/core/kind';
+import EnactPropTypes from '@enact/core/internal/prop-types';
 import PropTypes from 'prop-types';
 import React from 'react';
+
+import ForwardRef from '../ForwardRef';
 
 import componentCss from './BodyText.module.less';
 
@@ -19,7 +22,7 @@ import componentCss from './BodyText.module.less';
  * @ui
  * @public
  */
-const BodyText = kind({
+const BodyTextBase = kind({
 	name: 'ui:BodyText',
 
 	propTypes: /** @lends ui/BodyText.BodyText.prototype */ {
@@ -43,7 +46,14 @@ const BodyText = kind({
 		 * @default 'p'
 		 * @public
 		 */
-		component: PropTypes.oneOfType([PropTypes.string, PropTypes.func]),
+		component: EnactPropTypes.renderable,
+
+		/**
+		 * Called with a reference to [component]{@link ui/BodyText.BodyText#component}
+		 *
+		 * @private
+		 */
+		componentRef: PropTypes.func,
 
 		/**
 		 * Customizes the component by mapping the supplied collection of CSS class names to the
@@ -75,18 +85,22 @@ const BodyText = kind({
 		className: ({centered, styler}) => styler.append({centered})
 	},
 
-	render: ({component: Component, ...rest}) => {
+	render: ({component: Component, componentRef, ...rest}) => {
 		delete rest.centered;
 
 		return (
 			<Component
+				ref={componentRef}
 				{...rest}
 			/>
 		);
 	}
 });
 
+const BodyText = ForwardRef({prop: 'componentRef'}, BodyTextBase);
+
 export default BodyText;
 export {
-	BodyText
+	BodyText,
+	BodyTextBase
 };