+++ packages/ui/Repeater/Repeater.js
@@ -10,6 +10,8 @@ import EnactPropTypes from '@enact/core/internal/prop-types';
 import PropTypes from 'prop-types';
 import React from 'react';
 
+import ForwardRef from '../ForwardRef';
+
 /**
  * A stateless component that stamps out copies of `childComponent`.
  *
@@ -77,6 +79,14 @@ const RepeaterBase = kind({
 		component: EnactPropTypes.renderable,
 
 		/**
+		 * Called with a reference to [component]{@link ui/Repeater.Repeater#component}
+		 *
+		 * @type {Function}
+		 * @private
+		 */
+		componentRef: PropTypes.func,
+
+		/**
 		 * The property on each `childComponent` that receives the index of the item in the `Repeater`.
 		 *
 		 * @type {String}
@@ -116,15 +126,20 @@ const RepeaterBase = kind({
 		}
 	},
 
-	render: ({component: Component, ...rest}) => {
+	render: ({component: Component, componentRef, ...rest}) => {
 		delete rest.childComponent;
 		delete rest.childProp;
 		delete rest.indexProp;
 		delete rest.itemProps;
 
-		return <Component role="list" {...rest} />;
+		return <Component ref={componentRef} role="list" {...rest} />;
 	}
 });
 
-export default RepeaterBase;
-export {RepeaterBase as Repeater, RepeaterBase};
+const Repeater = ForwardRef({prop: 'componentRef'}, RepeaterBase);
+
+export default Repeater;
+export {
+	Repeater,
+	RepeaterBase
+};