+++ packages/ui/Spinner/Spinner.js
@@ -18,6 +18,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 
 import FloatingLayer from '../FloatingLayer';
+import ForwardRef from '../ForwardRef';
 
 import componentCss from './Spinner.module.less';
 
@@ -29,7 +30,7 @@ import componentCss from './Spinner.module.less';
  * @ui
  * @public
  */
-const Spinner = kind({
+const SpinnerBase = kind({
 	name: 'ui:Spinner',
 
 	propTypes: /** @lends ui/Spinner.Spinner.prototype */ {
@@ -73,6 +74,13 @@ const Spinner = kind({
 		 * @public
 		 */
 		centered: PropTypes.bool,
+
+		/**
+		 * Called with a reference to [component]{@link ui/Spinner.Spinner#component}
+		 *
+		 * @private
+		 */
+		componentRef: PropTypes.func,
 
 		/**
 		 * Customizes the component by mapping the supplied collection of CSS class names to the
@@ -139,7 +147,7 @@ const Spinner = kind({
 		)
 	},
 
-	render: ({blockClickOn, component: Component, scrimClassName, scrimType, spinnerContainerClassName, ...rest}) =>  {
+	render: ({blockClickOn, component: Component, componentRef, scrimClassName, scrimType, spinnerContainerClassName, ...rest}) =>  {
 		delete rest.centered;
 		delete rest.paused;
 		delete rest.scrim;
@@ -148,7 +156,7 @@ const Spinner = kind({
 			case 'screen': {
 				return (
 					<FloatingLayer noAutoDismiss open scrimType={scrimType}>
-						<Component {...rest} />
+						<Component ref={componentRef} {...rest} />
 					</FloatingLayer>
 				);
 			}
@@ -156,20 +164,23 @@ const Spinner = kind({
 				return (
 					<div className={spinnerContainerClassName}>
 						<div className={scrimClassName} />
-						<Component {...rest} />
+						<Component ref={componentRef} {...rest} />
 					</div>
 				);
 			}
 			default: {
 				return (
-					<Component {...rest} />
+					<Component ref={componentRef} {...rest} />
 				);
 			}
 		}
 	}
 });
 
+const Spinner = ForwardRef({prop: 'componentRef'}, SpinnerBase);
+
 export default Spinner;
 export {
-	Spinner
+	Spinner,
+	SpinnerBase
 };